requirements.txt
pip~=21.1.1
setuptools~=56.2.0
PyYAML~=5.4.1
PyQt5~=5.15.4